+++ modules/protocol/ptlink.c
@@ -123,7 +123,7 @@ static unsigned int ptlink_server_login(void)
 
 	sts("CAPAB :QS PTS4");
 	sts("SERVER %s 1 Atheme-%s :%s", me.name, version, me.desc);
-	sts("SVINFO 10 3 0 :%ld", CURRTIME);
+	sts("SVINFO 10 3 0 :%lu", (unsigned long)CURRTIME);
 
 	services_init();
 
@@ -135,7 +135,7 @@ static void ptlink_introduce_nick(user_t *u)
 {
 	const char *omode = is_ircop(u) ? "o" : "";
 
-	sts("NICK %s 1 %ld +i%sp %s %s %s %s :%s", u->nick, u->ts, omode, u->user, u->host, u->host, me.name, u->gecos);
+	sts("NICK %s 1 %lu +i%sp %s %s %s %s :%s", u->nick, (unsigned long)u->ts, omode, u->user, u->host, u->host, me.name, u->gecos);
 }
 
 /* invite a user to a channel */
@@ -162,11 +162,11 @@ static void ptlink_wallops_sts(const char *text)
 static void ptlink_join_sts(channel_t *c, user_t *u, boolean_t isnew, char *modes)
 {
 	if (isnew)
-		sts(":%s SJOIN %ld %s %s :@%s", me.name, c->ts, c->name,
-				modes, u->nick);
+		sts(":%s SJOIN %lu %s %s :@%s", me.name, (unsigned long)c->ts,
+				c->name, modes, u->nick);
 	else
-		sts(":%s SJOIN %ld %s + :@%s", me.name, c->ts, c->name,
-				u->nick);
+		sts(":%s SJOIN %lu %s + :@%s", me.name, (unsigned long)c->ts,
+				c->name, u->nick);
 }
 
 /* kicks a user from a channel */
@@ -265,7 +265,7 @@ static void ptlink_topic_sts(channel_t *c, const char *setter, time_t ts, time_t
 	if (!me.connected || !c)
 		return;
 
-	sts(":%s TOPIC %s %s %ld :%s", chansvs.nick, c->name, setter, ts, topic);
+	sts(":%s TOPIC %s %s %lu :%s", chansvs.nick, c->name, setter, (unsigned long)ts, topic);
 }
 
 /* mode wrapper */
@@ -447,14 +447,14 @@ static void m_sjoin(sourceinfo_t *si, int parc, char *parv[])
 			{
 				/* it's a service, reop */
 				sts(":%s PART %s :Reop", cu->user->nick, c->name);
-				sts(":%s SJOIN %ld %s + :@%s", me.name, ts, c->name, cu->user->nick);
+				sts(":%s SJOIN %lu %s + :@%s", me.name, (unsigned long)ts, c->name, cu->user->nick);
 				cu->modes = CMODE_OP;
 			}
 			else
 				cu->modes = 0;
 		}
 
-		slog(LG_DEBUG, "m_sjoin(): TS changed for %s (%ld -> %ld)", c->name, c->ts, ts);
+		slog(LG_DEBUG, "m_sjoin(): TS changed for %s (%lu -> %lu)", c->name, (unsigned long)c->ts, (unsigned long)ts);
 
 		c->ts = ts;
 		hook_call_event("channel_tschange", c);
